Update day 7 with a recursive FS buildup
My initial approach was to use a stack of folders and manage this stack while
the instructions are being red. A recursive approach appeared to be slightly
better-looking, hence it left this version afterwards. My initial implementation
was keeping track of the current folder using a mapping of dir path -> dir.
This approach was introduced as I mistakenly decided that I hit a cyclic symlink
in one of the folders, which appeared to be false. What happened is: i used
folder names as ref pointers, assuming these are unique. It is not the case
either. The new version removes visited checks that were used to guard against a
secondary node traversal.